Allow larger JS ArrayBuffer allocations in PartitionAlloc
Allow allocating larger JS ArrayBuffers directly without needing the
workaround of allocating a WebAssembly memory.

To support allocations up to ~16 GiB (16 GiB - kSuperPageSize) for
ArrayBuffers without adding runtime overhead for normal allocations, we
retain the normal constexpr bounds check against MaxAllocationSize() on
the fast path. Only if that check fails do we check whether the
allocation has AllocFlags::kAllowGigaAllocations and occurs in a
partition configured with PartitionOptions::allow_giga_allocations,
checking against the higher MaxGigaAllocationSize() limit only then.

TAG=agy
CONV=de69a7fe-0685-4882-85dc-5f4d8c93a314

Bug: 40055619
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8402423
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1704039}

+++ base/allocator/partition_allocator/src/partition_alloc/partition_bucket.h
@@ -37,7 +37,7 @@ struct PartitionBucket {
 
   SlotSpanMetadata* empty_slot_spans_head;
   SlotSpanMetadata* decommitted_slot_spans_head;
-  uint32_t slot_size;
+  size_t slot_size;
   uint32_t num_system_pages_per_slot_span
       : kPartitionNumSystemPagesPerSlotSpanBits;
   uint32_t num_full_slot_spans : 24;
@@ -67,7 +67,7 @@ struct PartitionBucket {
 
   // Public API.
   PA_COMPONENT_EXPORT(PARTITION_ALLOC)
-  void Init(uint32_t new_slot_size);
+  void Init(size_t new_slot_size);
 
   // Sets |is_already_zeroed| to true if the allocation was satisfied by
   // requesting (a) new page(s) from the operating system, or false otherwise.
